Move librosa functions to a separate file
